backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
anthropic==0.28.0
python-dotenv==1.0.1
